Feat: added new tshirt images, renamed folder and files and updates import files
