+++ src/components/ui/terms-and-conditions-section/terms-and-conditions-content.tsx
@@ -1,5 +1,5 @@
 import FooterSectionContent from "../footer-section/footer-content";
-import NavbarContent from "../nav-bar-section/nav-bar-content";
+import NavbarContent from "../navbar-section/navbar-content";
 
 export default function TermsAndConditionsContent(){
     return (
